main_project/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
celery
redis
fastapi-mail
python-jose
alembic
passlib
bcrypt